Make API DTO money fields decimal
- Transaction/CashVault/FinancialAnalytics API models: double -> decimal for Amount/Value/Limitation/Total* fields
- Keeps AutoMapper mapping to service models consistent (no implicit double<->decimal conversion)

+++ SmartWallet/Models/CashVault/CreateTransactionEndpointApiModel.cs
@@ -13,7 +13,7 @@ public class CreateTransactionEndpointApiModel
 	/// <summary>
 	/// Ограничение трат
 	/// </summary>
-	public double? Limitation { get; set; }
+	public decimal? Limitation { get; set; }
 	
 	/// <summary>
 	/// Флаг указывающий, что конечная точка является денежным хранилищем
